+++ src/components/channels/messages/Index.vue
@@ -7,7 +7,7 @@
       :items="messages"
       :date="from"
       :mode="mode"
-      :viewConfig="config"
+      :viewConfig="viewConfig"
       :colsConfigurator="'toolbar'"
       :i18n="i18n"
       :filter="filter"
@@ -35,7 +35,8 @@
 </template>
 
 <script>
-  import { VirtualScrollList } from 'qvirtualscroll'
+  import { VirtualScrollList, channelsMessagesModule } from 'qvirtualscroll'
+  import Vue from 'vue'
   import { date, Toast, LocalStorage } from 'quasar-framework'
   import filterMessages from '../../../mixins/filterMessages'
   import MessagesListItem from './MessagesListItem.vue'
@@ -46,38 +47,16 @@
       'date',
       'activeId',
       'limit',
-      'moduleName'
+      'config'
     ],
     data () {
       return {
         selectedItemKey: null,
-        theme: {
-          color: 'white',
-          bgColor: 'dark',
-          contentInverted: true,
-          controlsInverted: true
-        },
+        theme: this.config.theme,
         i18n: {},
-        config: {
-          needShowFilter: true,
-          needShowMode: false,
-          needShowPageScroll: 'right',
-          needShowDate: false
-        },
-        actions: [
-          {
-            icon: 'mdi-view-list',
-            label: 'view',
-            classes: '',
-            type: 'view'
-          },
-          {
-            icon: 'mdi-content-copy',
-            label: 'copy',
-            classes: '',
-            type: 'copy'
-          }
-        ]
+        viewConfig: this.config.viewConfig,
+        actions: this.config.actions,
+        moduleName: this.config.vuexModuleName
       }
     },
     computed: {
@@ -226,6 +205,7 @@
       }
     },
     async created () {
+      this.$store.registerModule(this.moduleName, channelsMessagesModule(this.$store, Vue))
       this.currentLimit = this.limit
       if (this.activeId) {
         this.$store.commit(`${this.moduleName}/setActive`, this.activeId)
@@ -237,6 +217,7 @@
     },
     destroyed () {
       this.$store.commit(`${this.moduleName}/clear`)
+      this.$store.unregisterModule(this.moduleName)
     },
     mixins: [filterMessages],
     components: { VirtualScrollList, MessagesListItem }