+++ src/components/devices/Index.vue
@@ -1,31 +1,5 @@
 <template>
   <div>
-    <q-toolbar slot="header" color="dark" v-if="active">
-      <q-item class="no-padding" style="max-width: 50%">
-        <q-item-main>
-          <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name || `#${selectedItem.id}`}}<q-tooltip v-if="$q.platform.is.desktop && selectedItem.name">{{selectedItem.name}}</q-tooltip></q-item-tile>
-          <q-item-tile sublabel style="font-size: 0.8rem" v-if="selectedItem.ident">{{selectedItem.ident}}</q-item-tile>
-        </q-item-main>
-        <q-item-side class="text-right"><q-icon color="white" size="2rem" name="mdi-menu-down" /></q-item-side>
-        <q-popover fit ref="popoverActive">
-          <q-list link separator class="scroll">
-            <q-item
-              v-for="(item, index) in items"
-              :key="index"
-              @click="active = item.id, $refs.popoverActive.close(), $emit('view-data-hide')"
-            >
-              <q-item-main>
-                <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: $q.platform.is.mobile ? '' : '140px'}">{{item.name || `#${item.id}`}}</q-item-tile>
-                <q-item-tile sublabel><small>{{item.ident}}</small></q-item-tile>
-              </q-item-main>
-              <q-item-side class="text-right" v-if="item.name"><small>#{{item.id.toString()}}</small></q-item-side>
-            </q-item>
-          </q-list>
-        </q-popover>
-      </q-item>
-      <q-btn flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'">{{$q.platform.is.mobile ? '' : modeModel ? 'Real-time' : 'History'}}</q-btn>
-      <q-icon v-if="isCustomer" size="1.5rem" style="position: absolute;right: 0;" class="on-left cursor-pointer" name="mdi-format-align-middle" @click="settingsClickHandler"/>
-    </q-toolbar>
     <div class="text-center" style="display: flex; justify-content: center; font-size: 1.5rem" v-if="!active">
       <div class="text-grey-3" style="margin-top: 50px">
         Select device
@@ -49,48 +23,72 @@
         </q-btn>
       </div>
     </div>
-    <logs
-      ref="logs"
-      :mode="mode"
-      :activeId="active"
-      originPattern="registry/devices/:id"
-      :isEnabled="!!+size[0]"
-      :delay="delay"
-      :moduleName="mainLogModuleName"
-      v-if="isCustomer && +size[0] && active"
-      :style="{minHeight: `calc(${size[0]}vh - ${+size[1] ? '50px' : '100px'})`, position: 'relative'}"
-      @view-log-message="viewLogMessagesHandler"
-    />
-    <messages
-      ref="messages"
-      @view-data="viewDataHandler"
-      :mode="mode"
-      :activeId="active"
-      :isEnabled="!!+size[1]"
-      :delay="delay"
-      :limit="limit"
-      :moduleName="mainChannelModuleName"
-      v-if="+size[1] && active"
-      :style="{minHeight: `calc(${size[1]}vh - ${+size[0] ? '50px' : '100px'})`, position: 'relative'}"
-    />
+    <template v-else>
+      <q-toolbar slot="header" color="dark">
+        <q-item class="no-padding" style="max-width: 50%">
+          <q-item-main>
+            <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: '140px'}">{{selectedItem.name || `#${selectedItem.id}`}}</q-item-tile>
+            <q-item-tile sublabel style="font-size: 0.8rem" v-if="selectedItem.ident">{{selectedItem.ident}}</q-item-tile>
+          </q-item-main>
+          <q-item-side class="text-right"><q-icon color="white" size="2rem" name="mdi-menu-down" /></q-item-side>
+          <q-popover fit ref="popoverActive">
+            <q-list link separator class="scroll">
+              <q-item
+                v-for="(item, index) in items"
+                :key="index"
+                @click="active = item.id, $refs.popoverActive.close(), $emit('view-data-hide')"
+              >
+                <q-item-main>
+                  <q-item-tile label class="ellipsis overflow-hidden" :style="{maxWidth: $q.platform.is.mobile ? '' : '140px'}">{{item.name || `#${item.id}`}}</q-item-tile>
+                  <q-item-tile sublabel><small>{{item.ident}}</small></q-item-tile>
+                </q-item-main>
+                <q-item-side class="text-right" v-if="item.name"><small>#{{item.id.toString()}}</small></q-item-side>
+              </q-item>
+            </q-list>
+          </q-popover>
+        </q-item>
+        <q-btn flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'">{{$q.platform.is.mobile ? '' : modeModel ? 'Real-time' : 'History'}}</q-btn>
+        <q-icon v-if="isCustomer" size="1.5rem" style="position: absolute;right: 0;" class="on-left cursor-pointer" name="mdi-format-align-middle" @click="settingsClickHandler"/>
+      </q-toolbar>
+      <logs
+        ref="logs"
+        :mode="mode"
+        :activeId="active"
+        originPattern="registry/devices/:id"
+        :isEnabled="!!+size[0]"
+        :delay="delay"
+        v-if="isCustomer && +size[0]"
+        :style="{minHeight: `calc(${size[0]}vh - ${+size[1] ? '50px' : '100px'})`, position: 'relative'}"
+        @view-log-message="viewLogMessagesHandler"
+        :config="config.logs"
+      />
+      <messages
+        ref="messages"
+        @view-data="viewDataHandler"
+        :mode="mode"
+        :activeId="active"
+        :isEnabled="!!+size[1]"
+        :delay="delay"
+        :limit="limit"
+        v-if="+size[1]"
+        :style="{minHeight: `calc(${size[1]}vh - ${+size[0] ? '50px' : '100px'})`, position: 'relative'}"
+        :config="config.messages"
+      />
+    </template>
   </div>
 </template>
 
 <script>
-  import Vue from 'vue'
   import { QToolbar, QSelect, QInput, Dialog, QIcon, QBtn, LocalStorage, QPopover, QList, QItem, QItemMain, QItemSide, QItemTile, QTooltip } from 'quasar-framework'
-  import { logsModule, devicesMessagesModule } from 'qvirtualscroll'
   import logs from '../logs/Index.vue'
   import messages from './messages/Index.vue'
-
-  const DEVICE_LOG_MODULE_NAME = 'devicesLogs',
-    MAIN_DEVICES_MODULE_NAME = 'devicesMessages'
 
   export default {
     props: [
       'limit',
       'delay',
-      'isCustomer'
+      'isCustomer',
+      'config'
     ],
     data () {
       let mode = LocalStorage.get.item('Toolbox-mode')
@@ -105,9 +103,7 @@
           {label: '50/50', value: '50/50'},
           {label: '40/60', value: '40/60'},
           {label: 'only messages', value: '0/100'}
-        ],
-        mainLogModuleName: DEVICE_LOG_MODULE_NAME,
-        mainChannelModuleName: MAIN_DEVICES_MODULE_NAME
+        ]
       }
     },
     computed: {
@@ -168,14 +164,6 @@
         this.$refs.messages.unselect()
       }
     },
-    beforeCreate () {
-      this.$store.registerModule(DEVICE_LOG_MODULE_NAME, logsModule(this.$store, Vue))
-      this.$store.registerModule(MAIN_DEVICES_MODULE_NAME, devicesMessagesModule(this.$store, Vue))
-    },
-    destroyed () {
-      this.$store.unregisterModule(DEVICE_LOG_MODULE_NAME)
-      this.$store.unregisterModule(MAIN_DEVICES_MODULE_NAME)
-    },
     created () {
       this.$store.dispatch('getItems', 'devices')
         .then(() => { this.$store.dispatch('getCustomer') })
@@ -211,6 +199,9 @@
           else if (this.isInit) {
             this.active = null
           }
+        }
+        else if (route.params && !route.params.id) {
+          this.active = null
         }
       },
       active (val) {