+++ src/components/logs/LogsListItem.vue
@@ -32,7 +32,7 @@
       etc () {
         let etcKeys = Object.keys(this.item).filter(key => !this.hasInCols(key))
         return etcKeys.reduce((acc, key) => {
-          if (key === 'event_origin' || key === 'event_text' || key === 'item_data' || key === 'error_text' || key === 'close_code' || key === 'http_data') { return acc }
+          if (key === 'event_origin' || key === 'event_text' || key === 'item_data' || key === 'error_text' || key === 'close_code' || key === 'http_data' || key === 'current' || key === 'updated' || key === 'error_code') { return acc }
           acc += `${key}: ${JSON.stringify(this.item[key])}; `
           return acc
         }, '') || '*Empty*'
@@ -176,9 +176,96 @@
             9: 'channel\'s parameters changed',
             10: 'connection closed upon user\'s request',
             11: 'terminal mode ended'
+          },
+          errorCodes = {
+            '-99': 'certificate error',
+            '-2': 'failed to establish TCP connection',
+            '-1': 'connection closed by flespi platform due to internal error',
+            1: 'unacceptable protocol version',
+            2: 'identifier rejected',
+            3: 'server unavailable',
+            4: 'bad username or password',
+            5: 'connection not authorized',
+            100: 'Continue',
+            101: 'Switching Protocols',
+            102: 'Processing',
+            200: 'OK',
+            201: 'Created',
+            202: 'Accepted',
+            203: 'Non-Authoritative Information',
+            204: 'No Content',
+            205: 'Reset Content',
+            206: 'Partial Content',
+            207: 'Multi-Status',
+            208: 'Already Reported',
+            226: 'IM Used',
+            300: 'Multiple Choices',
+            301: 'Moved Permanently',
+            302: 'Moved Temporarily',
+            303: 'See Other',
+            304: 'Not Modified',
+            305: 'Use Proxy',
+            307: 'Temporary Redirect',
+            308: 'Permanent Redirect',
+            400: 'Bad Request',
+            401: 'Unauthorized',
+            402: 'Payment Required',
+            403: 'Forbidden',
+            404: 'Not Found',
+            405: 'Method Not Allowed',
+            406: 'Not Acceptable',
+            407: 'Proxy Authentication Required',
+            408: 'Request Timeout',
+            409: 'Conflict',
+            410: 'Gone',
+            411: 'Length Required',
+            412: 'Precondition Failed',
+            413: 'Payload Too Large',
+            414: 'URI Too Long',
+            415: 'Unsupported Media Type',
+            416: 'Range Not Satisfiable',
+            417: 'Expectation Failed',
+            418: 'I’m a teapot',
+            422: 'Unprocessable Entity',
+            423: 'Locked',
+            424: 'Failed Dependency',
+            426: 'Upgrade Required',
+            428: 'Precondition Required',
+            429: 'Too Many Requests',
+            431: 'Request Header Fields Too Large',
+            449: 'Retry With',
+            451: 'Unavailable For Legal Reasons',
+            500: 'Internal Server Error',
+            501: 'Not Implemented',
+            502: 'Bad Gateway',
+            503: 'Service Unavailable',
+            504: 'Gateway Timeout',
+            505: 'HTTP Version Not Supported',
+            506: 'Variant Also Negotiates',
+            507: 'Insufficient Storage',
+            508: 'Loop Detected',
+            509: 'Bandwidth Limit Exceeded',
+            510: 'Not Extended',
+            511: 'Network Authentication Required',
+            520: 'Unknown Error',
+            521: 'Web Server Is Down',
+            522: 'Connection Timed Out',
+            523: 'Origin Is Unreachable',
+            524: 'A Timeout Occurred',
+            525: 'SSL Handshake Failed',
+            526: 'Invalid SSL Certificate'
           }
         let res = types[this.item.event_code] ? `${this.item.event_code}: ${types[this.item.event_code]}` : this.item.event_code
-        res += this.item.close_code ? `: ${closeCodes[this.item.close_code]}` : ''
+        res += this.item.close_code
+          ? ` (${closeCodes[this.item.close_code]})`
+          : (closeCodes[this.item.close_code]
+            ? `(${this.item.close_code})`
+            : '')
+        res += this.item.error_code
+          ? ` (${this.item.error_code}: ${errorCodes[this.item.error_code]})`
+          : (errorCodes[this.item.error_code]
+            ? `(${this.item.error_code})`
+            : '')
         return res
       }
     },